backend/rag-chatbot/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
qdrant-client==1.8.0
openai==1.10.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic-settings==2.1.0
python-multipart==0.0.6